unit1660.c drop redundant result cast
unit1658.c drop redundant result cast

unit1302.c drop redundant result casts

first.c drop redundant result cast

+++ tests/unit/unit1660.c
@@ -132,7 +132,7 @@ static CURLcode test_unit1660(const char *arg)
         continue;
       }
       else if(result) {
-        curl_mprintf("Input %d: error %d\n", i, (int)result);
+        curl_mprintf("Input %d: error %d\n", i, result);
         continue;
       }
     }
